chore: fix context for vite
